Use ssize_t instead of int in cases where variables are being assigned size_t values but we also need to worry about a potential value of -1.  Typically these are indices of loops that count down towards zero.  Although ssize_t is not in the C++ standard, normaliz already includes sys/types.h (which is the header that defines it), so this should be safe to use.

+++ source/libnormaliz/face_lattice.cpp
@@ -472,7 +472,7 @@ void FaceLattice<Integer>::compute(const long face_codim_bound, const bool verbo
     }
 
     size_t total_nr_faces = 0;
-    for (int i = prel_f_vector.size() - 1; i >= 0; --i) {
+    for (ssize_t i = prel_f_vector.size() - 1; i >= 0; --i) {
         if (prel_f_vector[i] != 0) {
             f_vector.push_back(prel_f_vector[i]);
             total_nr_faces += prel_f_vector[i];
